Move Islands upwards in DOM tree
