+++ frontend/src/client.entry.ts
@@ -1,6 +1,6 @@
 import Banner from "./components/Banner";
 import CountryStatsList from "./components/CountryStatsList";
-import DatawrapperLink from "./components/DatawrapperLink";
+import DatawrapperLink from "./components/DatawrapperLinkList";
 import Eyes from "./components/Eyes";
 import GroupStatsList from "./components/GroupStatsList";
 import MemberVotesList from "./components/MemberVotesList";
